metrics: Also ignore samples with unfired gpu fence
On some thirdparty devices it's possible to have a timestamp *available*
for the present fence before we have one for the GPU fence.

Just ignore these samples -- the timestamps will be filled in during the
next poll cycle, and we can guarantee that only the last sample in the
buffer may be in this state.

Bug: b/110939796
Change-Id: I2c58ba99d0468fc8345cd44ebbfeea8bfcf67b74